chore(GUIConstants): moved class in utils package
