+++ src/main/java/it/units/sdm/quoridor/GUI/view/dialogs/HelpDialogView.java
@@ -1,7 +1,7 @@
 package it.units.sdm.quoridor.GUI.view.dialogs;
 
 import it.units.sdm.quoridor.GUI.view.managers.BorderManager;
-import it.units.sdm.quoridor.GUI.GUIConstants;
+import it.units.sdm.quoridor.utils.GUIConstants;
 
 import javax.swing.*;
 import java.awt.*;
