+++ src/main/java/it/units/sdm/quoridor/GUI/view/dialogs/NotificationDialogView.java
@@ -1,6 +1,6 @@
 package it.units.sdm.quoridor.GUI.view.dialogs;
 
-import it.units.sdm.quoridor.GUI.GUIConstants;
+import it.units.sdm.quoridor.utils.GUIConstants;
 
 import javax.swing.*;
 import java.awt.*;
