+++ src/main/java/it/units/sdm/quoridor/GUI/view/dialogs/QuitDialogView.java
@@ -1,7 +1,7 @@
 package it.units.sdm.quoridor.GUI.view.dialogs;
 
 import it.units.sdm.quoridor.GUI.controller.GameController;
-import it.units.sdm.quoridor.GUI.GUIConstants;
+import it.units.sdm.quoridor.utils.GUIConstants;
 
 import javax.swing.*;
 import java.awt.*;
