+++ src/main/java/it/units/sdm/quoridor/GUI/view/panels/GameBoardPanelComponent.java
@@ -1,6 +1,6 @@
 package it.units.sdm.quoridor.GUI.view.panels;
 
-import it.units.sdm.quoridor.GUI.GUIConstants;
+import it.units.sdm.quoridor.utils.GUIConstants;
 import it.units.sdm.quoridor.GUI.controller.GameController;
 import it.units.sdm.quoridor.GUI.view.managers.BorderManager;
 import it.units.sdm.quoridor.model.AbstractPawn;
