+++ src/main/java/it/units/sdm/quoridor/GUI/view/panels/HelpQuitPanelComponent.java
@@ -1,7 +1,7 @@
 package it.units.sdm.quoridor.GUI.view.panels;
 
 import it.units.sdm.quoridor.GUI.view.managers.DialogManager;
-import it.units.sdm.quoridor.GUI.GUIConstants;
+import it.units.sdm.quoridor.utils.GUIConstants;
 
 import javax.swing.*;
 import java.awt.*;
