+++ src/main/java/it/units/sdm/quoridor/GUI/view/panels/RootPanelComponent.java
@@ -1,6 +1,6 @@
 package it.units.sdm.quoridor.GUI.view.panels;
 
-import it.units.sdm.quoridor.GUI.GUIConstants;
+import it.units.sdm.quoridor.utils.GUIConstants;
 import it.units.sdm.quoridor.GUI.view.managers.GameBoardPanelLayoutManager;
 import it.units.sdm.quoridor.GUI.controller.GameController;
 import it.units.sdm.quoridor.GUI.view.managers.PlayerPanelsManager;
