+++ src/main/java/it/units/sdm/quoridor/utils/GUIConstants.java
@@ -1,4 +1,4 @@
-package it.units.sdm.quoridor.GUI;
+package it.units.sdm.quoridor.utils;
 
 import java.awt.Color;
 import java.awt.Font;
